+++ robozonky-api/src/main/java/com/github/robozonky/api/Ratio.java
@@ -24,6 +24,8 @@ import java.util.Objects;
 
 import javax.json.bind.annotation.JsonbTypeAdapter;
 
+import com.github.robozonky.internal.remote.adapters.RatioAdapter;
+
 @JsonbTypeAdapter(RatioAdapter.class)
 public final class Ratio extends Number implements Comparable<Ratio> {
